chore: remove request folder
