+++ api/src/CrazyCards.Application/Core/Cards/Commands/CreateCardValidator.cs
