+++ api/src/CrazyCards.Application/Core/Classes/Commands/CreateClassHandler.cs
