+++ api/src/CrazyCards.Application/Core/Classes/Commands/CreateClassValidator.cs
