+++ api/src/CrazyCards.Application/Core/Images/Commands/CreateImageCommand.cs
@@ -2,7 +2,7 @@
 using CrazyCards.Application.Contracts.Images;
 using CrazyCards.Domain.Primitives.Result;
 
-namespace CrazyCards.Application.Core.Images.Commands.CreateImage;
+namespace CrazyCards.Application.Core.Images.Commands;
 
 public record CreateImageCommand(Stream Stream, string MimeType)
     : ICommand<Result<ImageResponse>>;
