+++ api/src/CrazyCards.Application/Core/Player/Commands/CreatePlayerCommand.cs
@@ -2,6 +2,6 @@
 using CrazyCards.Application.Contracts.Players;
 using CrazyCards.Domain.Primitives.Result;
 
-namespace CrazyCards.Application.Core.Player.Commands.CreatePlayer;
+namespace CrazyCards.Application.Core.Player.Commands;
 
 public record CreatePlayerCommand(string Username, string Email, string Password) : ICommand<Result<PlayerResponse>>;
