+++ api/src/CrazyCards.Application/Core/Player/Commands/CreatePlayerValidator.cs
