+++ api/src/CrazyCards.Application/Core/Skin/Commands/CreateSkinCommand.cs
@@ -2,6 +2,6 @@
 using CrazyCards.Application.Contracts.Skin;
 using CrazyCards.Domain.Primitives.Result;
 
-namespace CrazyCards.Application.Core.Skin.Commands.CreateSkin;
+namespace CrazyCards.Application.Core.Skin.Commands;
 
 public record CreateSkinCommand(string Name, Stream Stream, string MimeType) : ICommand<Result<SkinResponse>>;
