+++ api/src/CrazyCards.Application/Core/Skin/Queries/GetSkinByIdQuery.cs
